Separate service interfaces and implementations into dedicated folders
Reorganized the Services directory structure to improve code organization:
- Created Services/Interfaces folder for all service interfaces (I*.cs)
- Created Services/Implementations folder for all service implementations
- Updated all namespaces:
  - GLMod.Services.Interfaces for interface files
  - GLMod.Services.Implementations for implementation files
- Updated using statements in GLMod.cs and MainMenuManagerPatch.cs

This separation provides better separation of concerns and makes the
architecture more maintainable.
