+++ GLMod/GLMod.cs
@@ -9,7 +9,8 @@ using System.IO;
 using System.Linq;
 using GLMod.Enums;
 using GLMod.Constants;
-using GLMod.Services;
+using GLMod.Services.Interfaces;
+using GLMod.Services.Implementations;
 using Random = System.Random;
 using GLMod.GLEntities;
 using GLMod.Class;
