+++ GLMod/Services/Interfaces/IAuthenticationService.cs
@@ -1,6 +1,6 @@
 using System.Collections;
 
-namespace GLMod.Services
+namespace GLMod.Services.Interfaces
 {
     /// <summary>
     /// Interface for authentication and user session management
