+++ GLMod/Services/Interfaces/IIntegrityService.cs
@@ -1,6 +1,6 @@
 using System.Collections;
 
-namespace GLMod.Services
+namespace GLMod.Services.Interfaces
 {
     /// <summary>
     /// Interface for managing file integrity verification
