+++ GLMod/Services/Interfaces/IItemService.cs
@@ -2,7 +2,7 @@ using GLMod.GLEntities;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace GLMod.Services
+namespace GLMod.Services.Interfaces
 {
     /// <summary>
     /// Interface for managing items and DLC ownership
